+++ docs/src/changelog.md
@@ -8,10 +8,12 @@ Release notes:
 
 ### Added
 - [`as_section`](@ref) for assembling 'ctd' data into a section.
-- [`get_section`](@ref) for downloading oceanographic 'section' data.
-- [`read_section`](@ref) for reading 'section' data.
+- [`get_section`](@ref) for downloading oceanographic section data.
+- [`read_section`](@ref) for reading Section data.
+- [`grid_section`](@ref) for gridding Section data.
 - [`plot_section`](@ref) for plotting oceanographic 'section' data.
 - [`read_ctd_woce`](@ref) for reading CTD data in WOCE format (used by many data servers)
+- [`grid_ctd`](@ref) for gridding Ctd data.
 
 
 # Changelog from version 0.2.0 to version 0.2.1